修改数据类型

+++ src/com/company/ui/MainApplication.java
@@ -12,7 +12,7 @@ public class MainApplication {
 
     public static People user;
     public static OriginalLoginPage app;
-    public static int num=0;
+    public static short num=0;
 
 
     public static void main(String[] args){
